Broke out main base shear app component into sub components
