+++ src/components/features/BaseShearApp/Header.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const Header = () => (
+  <div className="flex justify-center items-center">
+    <h1 className="text-center text-6xl text-black font-bold mb-4">Base Shear Diagram</h1>
+  </div>
+);
+
+export default Header;
